requirements.txt
flask
flask-cors
ytmusicapi
eyed3
pyjwt
oauthlib
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
platformdirs
waitress
virtualenv